Skip transactions if already in book
